+++ sdk/src/runtime/lib/db/SqliteDurableObject.ts
@@ -1,20 +1,13 @@
-import {
-  Kysely,
-  Driver,
-  DatabaseConnection,
-  CompiledQuery,
-  QueryResult,
-  SqliteAdapter,
-  SqliteIntrospector,
-  SqliteQueryCompiler,
-} from "kysely";
 import { DODialect } from "kysely-do";
 import { DurableObject } from "cloudflare:workers";
-import { createMigrator } from "./migrations.js";
+
+import { Kysely, CompiledQuery, QueryResult } from "kysely";
+import { createMigrator } from "./index.js";
 import debug from "../debug.js";
 
 const log = debug("sdk:do-db");
 
+// Base class for Durable Objects that need Kysely database access
 export class SqliteDurableObject<T = any> extends DurableObject {
   public migrations: Record<string, any>;
   public kysely: Kysely<T>;
@@ -74,107 +67,5 @@ export class SqliteDurableObject<T = any> extends DurableObject {
       queryId: {} as any,
     } as CompiledQuery<unknown>);
     return result as QueryResult<R>;
-  }
-}
-
-// DOWorkerDialect - mimics DODialect but proxies to DO
-export class DOWorkerDialect {
-  #config: { stub: any };
-
-  constructor(config: { stub: any }) {
-    this.#config = config;
-  }
-
-  createAdapter() {
-    return new SqliteAdapter();
-  }
-
-  createDriver() {
-    return new DOWorkerDriver(this.#config);
-  }
-
-  createQueryCompiler() {
-    return new SqliteQueryCompiler();
-  }
-
-  createIntrospector(db: any) {
-    return new SqliteIntrospector(db);
-  }
-}
-
-class DOWorkerDriver implements Driver {
-  #config: { stub: any };
-
-  constructor(config: { stub: any }) {
-    this.#config = config;
-  }
-
-  async init() {}
-
-  async acquireConnection(): Promise<DatabaseConnection> {
-    return new DOWorkerConnection(this.#config.stub.kyselyExecuteQuery);
-  }
-
-  async beginTransaction(conn: any) {
-    return await conn.beginTransaction();
-  }
-
-  async commitTransaction(conn: any) {
-    return await conn.commitTransaction();
-  }
-
-  async rollbackTransaction(conn: any) {
-    return await conn.rollbackTransaction();
-  }
-
-  async releaseConnection(_conn: any) {}
-
-  async destroy() {}
-}
-
-class DOWorkerConnection implements DatabaseConnection {
-  #executeQuery: (compiledQuery: {
-    sql: string;
-    parameters: readonly unknown[];
-  }) => Promise<QueryResult<any>>;
-
-  constructor(
-    executeQuery: (compiledQuery: {
-      sql: string;
-      parameters: readonly unknown[];
-    }) => Promise<QueryResult<any>>,
-  ) {
-    this.#executeQuery = executeQuery;
-  }
-
-  async executeQuery<R>(compiledQuery: {
-    sql: string;
-    parameters: readonly unknown[];
-  }): Promise<QueryResult<R>> {
-    log("Forwarding query to Durable Object: %s", compiledQuery.sql);
-
-    // Call the DO's kyselyExecuteQuery method
-    const result = await this.#executeQuery({
-      sql: compiledQuery.sql,
-      parameters: compiledQuery.parameters,
-    });
-
-    return result as QueryResult<R>;
-  }
-
-  async beginTransaction() {
-    throw new Error("Transactions are not supported yet.");
-  }
-
-  async commitTransaction() {
-    throw new Error("Transactions are not supported yet.");
-  }
-
-  async rollbackTransaction() {
-    throw new Error("Transactions are not supported yet.");
-  }
-
-  async *streamQuery(_compiledQuery: any, _chunkSize?: any) {
-    throw new Error("DO Driver does not support streaming");
   }
 }
