+++ sdk/src/runtime/lib/db/index.ts
@@ -1,3 +1,3 @@
 export * from "./migrations.js";
-export * from "./createDb.js";
 export * from "./SqliteDurableObject.js";
+export * from "./createDb.js";
